test/kyu7/MumblingTest.java
package kyu7;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

class MumblingTest {

    @ParameterizedTest
    @MethodSource
    public void testing(String actual, String expected) {
        assertEquals(expected, Mumbling.accum(actual));
    }
    @ParameterizedTest
    @MethodSource("testing")
    public void testingClever(String actual, String expected) {
        assertEquals(expected, Mumbling.accumClever(actual));
    }

    public static Stream<Arguments> testing() {
        return Stream.of(
                Arguments.of("ZpglnRxqenU", "Z-Pp-Ggg-Llll-Nnnnn-Rrrrrr-Xxxxxxx-Qqqqqqqq-Eeeeeeeee-Nnnnnnnnnn-Uuuuuuuuuuu"),
                Arguments.of("NyffsGeyylB", "N-Yy-Fff-Ffff-Sssss-Gggggg-Eeeeeee-Yyyyyyyy-Yyyyyyyyy-Llllllllll-Bbbbbbbbbbb"),
                Arguments.of("MjtkuBovqrU", "M-Jj-Ttt-Kkkk-Uuuuu-Bbbbbb-Ooooooo-Vvvvvvvv-Qqqqqqqqq-Rrrrrrrrrr-Uuuuuuuuuuu"),
                Arguments.of("EvidjUnokmM", "E-Vv-Iii-Dddd-Jjjjj-Uuuuuu-Nnnnnnn-Oooooooo-Kkkkkkkkk-Mmmmmmmmmm-Mmmmmmmmmmm"),
                Arguments.of("HbideVbxncC", "H-Bb-Iii-Dddd-Eeeee-Vvvvvv-Bbbbbbb-Xxxxxxxx-Nnnnnnnnn-Cccccccccc-Ccccccccccc"),
                Arguments.of("H", "H")
        );
    }

}